+++ crates/orchestrator/src/jobs/constants.rs
@@ -10,4 +10,5 @@ pub const JOB_METADATA_SNOS_BLOCK: &str = "block_number_to_run";
 pub const JOB_METADATA_SNOS_FACT: &str = "snos_fact";
 pub const JOB_METADATA_FAILURE_REASON: &str = "failure_reason";
 pub const JOB_METADATA_ERROR: &str = "error";
+pub const JOB_METADATA_PROCESSING_STARTED_AT: &str = "processing_started_at";
 pub const JOB_METADATA_PROCESSING_COMPLETED_AT: &str = "processing_completed_at";