+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Limits on SNOS job concurrency
 - Unique Id for running orchestrator
+- Added JOB_METADATA_PROCESSING_STARTED_AT
 - Added retry job endpoint for failed jobs
 - readme: setup instructions added
 - Added : Grafana dashboard
@@ -89,6 +90,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Fixed
 
+- fixed JOB_METADATA_PROCESSING_COMPLETED_AT usage
 - add jemallocator to fix memory leak
 - refactor: instrumentation
 - `is_worker_enabled` status check moved from `VerificationFailed` to `Failed`